refactor: refactor naming, remove route for error boundary
