+++ project-management-app/src/pages/SignUpPage/index.tsx
@@ -4,7 +4,7 @@ import { Navigate } from 'react-router-dom';
 import MainPaper from 'components/MainPaper';
 import SignUpForm from './components/SignUpForm';
 
-import { PATH } from 'components/Router/types';
+import { PATH } from 'components/AppRoutes/types';
 
 import styles from './SignUpPage.module.scss';
